--- src/link_kinesis_stream_to_firehose/config.yml
@@ -1,4 +1,0 @@
-name: 'link_kinesis_stream_to_firehose'
-runtime: 'python3.6'
-s3_bucket: 'nicor-dev'
-s3_key: 'deployments/lambdas/link_kinesis_stream_to_firehose.zip'
